requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.127.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
phonenumbers==9.0.21
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
python-multipart==0.0.21
qrcode==8.2
requests==2.32.5
starlette==0.50.0
stripe==14.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
